+++ src/test/java/racingcar/model/RacingGameTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.model;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
